requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.7
colorama>=0.4.6
tqdm>=4.66.1